Moved factory initialization into 'object_factory.py'
